Small update
Short update that provides a bit more explanation for what each form tab will contain. It also adds an additional div to the right side of the page for the model output to be displayed. A dummy graph from chart.js is displayed on the front end as a quick example.

This will need to be updated further so things display at the proper times (we have sent the csv to the server, a graph should be displayed in response.)