Error handling
Adjusted files post merge to properly display results after uploading and merging csv files. I also added an additional way to verify that the submitted files are of the correct extension, and return an error if they are not.

Finally, I added a 'catch all' error handling to zip_unpack, in case there is an unknown error in our process, that prevents the server from crashing.