Assign Column Names
If the file(s) uploaded by the user do not appear to contain a header, the program will generate column names based on the number of columns.

The program does assume that col 0 is the index and will set that accordingly.

Also removed code that prevented users from selecting new columns to graph.