Fixed csv/zip uploading functions
- Adjusted csv_upload to mirror how we handle csv files in zip_unpack.
- Removed code that would save uploaded zip files to root directory.
- Removed unnecessary upload folder.
- Specified decoding type for uploaded files.
- Added hidden directory checks when uploading zip folders.
- Removed unneeded imports.

A note for the future: Occasionally, the uploaded csv will be filled with mostly NaN values (including in the header). A check should be added to ensure NaN values in the header should default to generating column names. Another check for automatic indexing would also be good to add in.